reworking citations

+++ src/Agile.Chat/Agile.Chat.Domain/ChatThreads/ValueObjects/ChatThreadFIlterOptions.cs
@@ -3,6 +3,6 @@
 public class ChatThreadFilterOptions
 {
     public int DocumentLimit { get; set; } = 5;
-    public double? Strictness { get; set; }
+    public int? Strictness { get; set; }
     public List<string> Folders { get; set; } = new();
 }
